+++ Raptor/Graphics/ITexture.cs
@@ -5,21 +5,17 @@
 namespace Raptor.Graphics
 {
     using System;
+    using Raptor.Content;
 
     /// <summary>
     /// The texture to render to a screen.
     /// </summary>
-    public interface ITexture : IDisposable
+    public interface ITexture : IContent, IDisposable
     {
         /// <summary>
         /// Gets he ID of the texture.
         /// </summary>
         uint ID { get; }
-
-        /// <summary>
-        /// Gets the name of the texture.
-        /// </summary>
-        string Name { get; }
 
         /// <summary>
         /// Gets the width of the texture.