+++ Testing/RaptorTests/Fakes/IInvalidContent.cs
@@ -1,0 +1,15 @@
+﻿// <copyright file="IInvalidContent.cs" company="KinsonDigital">
+// Copyright (c) KinsonDigital. All rights reserved.
+// </copyright>
+
+namespace RaptorTests.Fakes
+{
+    using Raptor.Content;
+
+    /// <summary>
+    /// Used to represent invalid content for testing the <see cref="ContentLoader"/> class.
+    /// </summary>
+    public interface IInvalidContent : IContent
+    {
+    }
+}